refactor: Cards 및 Deck 객체 분리 및 테스트 이동
