+++ src/main/java/domain/card/Cards.java
@@ -2,10 +2,8 @@ package domain.card;
 
 import static domain.Constant.DELIMITER;
 
-import domain.ExceptionMessage;
 import domain.Rank;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -20,27 +18,12 @@ public class Cards {
         cards.add(card);
     }
 
-    public Card pull() {
-        validateIsEmpty();
-        return cards.removeFirst();
-    }
-
-    private void validateIsEmpty() {
-        if (cards.isEmpty()) {
-            throw new IllegalArgumentException(ExceptionMessage.EMPTY_CARDS.getMessage());
-        }
-    }
-
     public Card peek() {
         return cards.getFirst();
     }
 
     public int size() {
         return cards.size();
-    }
-
-    public void shuffle() {
-        Collections.shuffle(cards);
     }
 
     public int getTotalSum() {
