+++ src/main/java/domain/card/Deck.java
@@ -1,29 +1,49 @@
 package domain.card;
 
+import static domain.Constant.DELIMITER;
+
+import domain.ExceptionMessage;
 import domain.Rank;
 import domain.Suit;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class Deck {
-    private final Cards cards;
+    private final List<Card> deck;
 
     public Deck() {
-        this.cards = new Cards();
+        this.deck = new ArrayList<Card>();
 
         for (Suit suit : Suit.values()) {
             for (Rank rank : Rank.values()) {
-                cards.add(new Card(suit, rank));
+                deck.add(new Card(suit, rank));
             }
         }
 
-        cards.shuffle();
+        shuffle();
+    }
+
+    public void shuffle() {
+        Collections.shuffle(deck);
     }
 
     public Card pull() {
-        return cards.pull();
+        validateIsEmpty();
+        return deck.removeFirst();
+    }
+
+    private void validateIsEmpty() {
+        if (deck.isEmpty()) {
+            throw new IllegalArgumentException(ExceptionMessage.EMPTY_CARDS.getMessage());
+        }
     }
 
     @Override
     public String toString() {
-        return cards.toString();
+        return deck.stream()
+                .map(card -> card.toString())
+                .collect(Collectors.joining(DELIMITER));
     }
 }
