+++ src/test/java/domain/card/CardsTest.java
@@ -59,11 +59,4 @@ public class CardsTest {
         //then
         Assertions.assertEquals(cards.getTotalSum(), 13);
     }
-
-    @Test
-    void 뽑을_수_있는_카드가_존재하지_않으면_예외가_발생한다() {
-        Cards cards = new Cards();
-        org.assertj.core.api.Assertions.assertThatThrownBy(() -> cards.pull())
-                .isInstanceOf(IllegalArgumentException.class);
-    }
 }
