+++ src/test/java/domain/card/DeckTest.java
@@ -9,4 +9,15 @@ public class DeckTest {
     void 덱_생성_시_정상적으로_생성되어야_한다() {
         assertDoesNotThrow(() -> new Deck());
     }
+
+    @Test
+    void 뽑을_수_있는_카드가_존재하지_않으면_예외가_발생한다() {
+        Deck deck = new Deck();
+        for (int i = 0; i < 52; i++) {
+            deck.pull();
+        }
+
+        org.assertj.core.api.Assertions.assertThatThrownBy(() -> deck.pull())
+                .isInstanceOf(IllegalArgumentException.class);
+    }
 }
